typo float/double for the SIMD register

+++ libv/src/cyme_ed2bd4hkth1d4.cpp
@@ -50,9 +50,9 @@ namespace cyme {
     };
 } // end namespace
 
-    typedef double v8float __attribute((vector_size(32)));
+    typedef float v8float __attribute((vector_size(32)));
     typedef double v4double __attribute((vector_size(32)));
-    typedef double v4float __attribute((vector_size(16)));
+    typedef float v4float __attribute((vector_size(16)));
     typedef double v2double __attribute((vector_size(16)));
 
     v4double v4dexp(v4double a){
